requirements.txt
alembic==1.16.4
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
audioop-lts==0.2.2
audioread==3.1.0
blinker==1.9.0
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.5
fpdf==1.7.2
fpdf2==2.8.3
fsspec==2025.7.0
future==1.0.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==1.100.2
openai-whisper==20250625
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.5.0
playwright==1.54.0
pooch==1.8.2
proglog==0.1.12
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
regex==2024.11.6
requests==2.32.4
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.41
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.9.0
torchaudio==2.9.0+cpu
torchvision==0.2.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
